Change: Rename and move module for LineChart
Use the same name for the module as the component within and move it to
the location of the other base charts.
